requirements.txt
torch>=1.11.0
torchvision
numpy
tqdm
Pillow
facenet-pytorch
